move commands to separate files
