+++ packages/expo-type-information/src/commands/generateTypeFilesCommand.ts
@@ -1,0 +1,52 @@
+import commander from 'commander';
+import {
+  addCommonOptions,
+  getFileTypeInformationFromArgs,
+  parseCommandArguments,
+  runCommandOnWatch,
+  TypeInformationCommandCommonAllArguments,
+  writeToStableFile,
+} from './commandUtils';
+import { generateFullTsInterface } from '../typescriptGeneration';
+import path from 'path';
+
+export function generateTypeFilesCommand(cli: commander.Command) {
+  return addCommonOptions(cli.command('generate-type-files')).action(
+    async (options: TypeInformationCommandCommonAllArguments) => {
+      const parsedArgs = await parseCommandArguments(options, false);
+      if (!parsedArgs) {
+        return;
+      }
+      const { realInputPaths, realOutputPath } = parsedArgs;
+
+      const command = async () => {
+        const typeInfo = await getFileTypeInformationFromArgs(parsedArgs);
+        if (!typeInfo) {
+          return;
+        }
+        const generatedFiles = await generateFullTsInterface(typeInfo);
+        if (!generatedFiles) {
+          return;
+        }
+        const { moduleTypesFile, moduleViewsFiles, moduleNativeFile, indexFile } = generatedFiles;
+
+        const dirName = realOutputPath ?? path.dirname(realInputPaths[0] as string);
+        const writeFilePromises = [];
+        for (const outputFile of [
+          moduleTypesFile,
+          ...moduleViewsFiles,
+          moduleNativeFile,
+          indexFile,
+        ]) {
+          const outputFilePath = path.resolve(dirName, outputFile.name);
+          writeFilePromises.push(
+            writeToStableFile({ filePath: outputFilePath, content: outputFile.content })
+          );
+        }
+        await Promise.all(writeFilePromises);
+      };
+
+      runCommandOnWatch(parsedArgs, command);
+    }
+  );
+}
